DEVHUB-663: Adds new template
